Error Torelant and Card Size ยังไม่เสร็จ

+++ src/app/utils/BannerPage/useBannerPage.tsx
@@ -8,6 +8,8 @@ import { createBanner, deleteBanner } from "@/app/api/banner/route";
 import { useRouter } from "next/navigation";
 import ITryToastNotification from "@/app/components/Toast/ToastNotification";
 import { mutate } from "swr";
+import Swal from 'sweetalert2';
+
 
 export default function useBannerPage() {
     const schema = yup.object().shape({
@@ -22,49 +24,57 @@ export default function useBannerPage() {
         try {
             const image = data.image;
             const bannerUrl: any = await uploadFileToS3(image); // อัปโหลดภาพไปยัง Amazon S3 และรับ URL กลับมา
-    
             const result = await createBanner(bannerUrl);
-            
-
-            await ITryToastNotification({
-                type: "success",
-                text: "เพิ่มแบนเนอร์สำเร็จ"
-            })
-
+            Swal.fire({
+                icon: "success",
+                text: "Banner added successfully",
+                showConfirmButton: false,
+                timer: 1500
+            });
             mutate('getBanner');
-
             setValue("image", undefined)
-    
             console.log("Data saved to DynamoDB successfully result:", result);
         } catch (error) {
-            await ITryToastNotification({
-                type: "error",
-                text: "เพิ่มแบนเนอร์ล้มเหลว"
-            })
+            Swal.fire({
+                icon: "error",
+                title: "Oops...",
+                text: "Failed to add banner!",
+            });
             console.error("Error saving data to DynamoDB:", error);
         }
     }
 
     const onDeleteBanner = async (bannerId: string) => {
         try {
-            const result = await deleteBanner(bannerId);
+            Swal.fire({
+                title: "Are you sure?",
+                text: "Do you want to delete a banner, right?",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes, delete it!"
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    Swal.fire({
+                        icon: "success",
+                        text: "Banner delete successfully",
+                    });
+                    const result = await deleteBanner(bannerId);
+                    mutate('getBanner');
+                    console.log(`Banner with ID ${bannerId} deleted successfully result:`, result);
+                }
+            });
 
-            await ITryToastNotification({
-                type: "success",
-                text: "ลบแบนเนอร์สำเร็จ"
-            })
-    
-            mutate('getBanner');
-
-            console.log(`Banner with ID ${bannerId} deleted successfully result:`, result);
         } catch (error) {
-            await ITryToastNotification({
-                type: "error",
-                text: "ลบแบนเนอร์ล้มเหลว"
-            })
+            Swal.fire({
+                icon: "error",
+                title: "Oops...",
+                text: "Failed to delete banner!",
+            });
             console.log("Error deleting banner:", error);
         }
     }
-    
+
     return { register, setValue, watch, handleSubmit, onSubmit, onDeleteBanner }
 }
